Fix pulumi tags to not contain ",omitempty"
The Pulumi marshaling machinery doesn't seem to like when we
stuff additional info into the `pulumi:".."` tags. So we will
instead use a separate `pschema:".."` tag when we need to
specify anything opther than the name. In general, it's fine
to leave off the ",omitempty" parts anyway since we're using
pointers which will be unmarshaled as nil if missing anyway.